Order with details

+++ backEnd/dto/dto_detail.go
@@ -4,7 +4,7 @@ type OrderDetailDto struct {
 	DetailId   int     `json:"detail_id"`
 	Quantity   int     `json:"quantity"`
 	Price      float32 `json:"price"`
-	TotalPrice int     `json:"total_price"`
+	TotalPrice float32 `json:"total_price"`
 	ProductId  int     `json:"product_id"`
 	OrderId    int     `json:"order_id"`
 }
